extract items into lib folder
