+++ src/lib/instructors.ts
@@ -1,5 +1,7 @@
 import {request} from 'graphql-request'
+import {sanityClient} from 'utils/sanity-client'
 import groq from 'groq'
+import {stephanieEcklesQuery} from 'components/search/instructors/stephanie-eckles'
 import config from './config'
 
 export type Instructor = {
@@ -36,4 +38,27 @@ export async function loadInstructor(slug: string) {
   const {instructor} = await request(config.graphQLEndpoint, query, {slug})
 
   return instructor
+}
+
+const sanityInstructorHash = {
+  'stephanie-eckles': stephanieEcklesQuery,
+}
+
+type SelectedInstructor = keyof typeof sanityInstructorHash
+
+const canLoadSanityInstructor = (
+  selectedInstructor: string,
+): selectedInstructor is SelectedInstructor => {
+  const keyNames = Object.keys(sanityInstructorHash)
+
+  return keyNames.includes(selectedInstructor)
+}
+
+export const loadSanityInstructor = async (selectedInstructor: string) => {
+  if (!canLoadSanityInstructor(selectedInstructor)) return
+
+  const query = sanityInstructorHash[selectedInstructor]
+  if (!query) return
+
+  return await sanityClient.fetch(query)
 }
